Se cambian a double los valores de los promedios

+++ materia.c
@@ -153,12 +153,12 @@ void borrarMateriaPorNombre(Materia *listaNombre, char *nombre) {
 }
 
 void obtenerPromedioDeAprobados(Materia *materia){
-    float promedioDeAprobados = (materia->cantDeAprobados*100)*1.0/materia->cantDeEstudiantesRendieron;
+    double promedioDeAprobados = (materia->cantDeAprobados*100)*1.0/materia->cantDeEstudiantesRendieron;
     printf("El promedio de aprobacion es del %.2f\n", promedioDeAprobados);
 }
 
 void obtenerNotaPromedio(Materia *materia){
-    float notaPromedio = (materia->notas*100)*1.0/materia->cantDeEstudiantesRendieron;
+    double notaPromedio = (materia->notas*100)*1.0/materia->cantDeEstudiantesRendieron;
     printf("La nota promedio en es %.2f\n", notaPromedio);
 }
